edit mistakes after consultation
