+++ js/main.js
@@ -1,5 +1,6 @@
-import './map.js';
-import './pins.js';
+import {CARDS_COUNT, makeCards} from './make-cards.js';
+import {map} from './map.js';
+import {renderPins} from './pins.js';
 
-
-
+const mockObjects = makeCards(CARDS_COUNT);
+renderPins(map, mockObjects);
