+++ js/pins.js
@@ -1,50 +1,10 @@
 /* global L:readonly */
-import {CARDS_COUNT, makeCards} from './make-cards.js';
 import {createCardLayout} from './create-card-layout.js';
-import {map, TOKYO_CENTER, inputAddress} from './map.js';
 
-const MAIN_PIN_PARAMETERS = {
-  x: 50,
-  y: 82,
+const AdPinParameters = {
+  X: 25,
+  Y: 41,
 };
-
-const AD_PIN_PARAMETERS = {
-  x: 25,
-  y: 41,
-};
-
-const ADDRESS_DIGITS_AFTER_DECIMAL = 5;
-
-// Ставим главный пин на карту
-
-const mainIcon = L.icon({
-  iconUrl: 'leaflet/images/marker-icon-2x.png',
-  iconSize: [MAIN_PIN_PARAMETERS.x, MAIN_PIN_PARAMETERS.y],
-  iconAnchor: [(MAIN_PIN_PARAMETERS.x) / 2, MAIN_PIN_PARAMETERS.y],
-});
-
-const mainPin = L.marker({
-  lat: TOKYO_CENTER.x,
-  lng: TOKYO_CENTER.y,
-},
-{
-  draggable: true,
-  icon: mainIcon,
-},
-);
-
-mainPin.addTo(map);
-
-// При перемещении главного пина меняется значение поля ввода адреса
-
-mainPin.on('moveend', (evt) => {
-  inputAddress.value = `${evt.target.getLatLng().lat.toFixed(ADDRESS_DIGITS_AFTER_DECIMAL)},
-  ${evt.target.getLatLng().lng.toFixed(ADDRESS_DIGITS_AFTER_DECIMAL)}`;
-});
-
-// Ставим пины моков на карту и создаем для каждого попап
-
-const mockObjects = makeCards(CARDS_COUNT);
 
 const createPopupCard = (object) => {
   const popupElement = createCardLayout(object);
@@ -53,27 +13,32 @@ const createPopupCard = (object) => {
   return popupElement;
 }
 
-mockObjects.forEach((item) => {
-  const icon = L.icon({
-    iconUrl: 'leaflet/images/marker-icon.png',
-    iconSize: [AD_PIN_PARAMETERS.x, AD_PIN_PARAMETERS.y],
-    iconAnchor: [(AD_PIN_PARAMETERS.x) / 2, AD_PIN_PARAMETERS.y],
-  });
+const renderPins = function (map, array) {
+  array.forEach((item) => {
+    const icon = L.icon({
+      iconUrl: 'leaflet/images/marker-icon.png',
+      iconSize: [AdPinParameters.X, AdPinParameters.Y],
+      iconAnchor: [(AdPinParameters.X) / 2, AdPinParameters.Y],
+    });
 
-  const adPin = L.marker({
-    lat: item.location.x,
-    lng: item.location.y,
-  },
-  {
-    icon,
-  },
-  );
-
-  adPin
-    .addTo(map)
-    .bindPopup(createPopupCard(item),
-      {
-        keepInView: true,
-      },
+    const adPin = L.marker({
+      lat: item.location.x,
+      lng: item.location.y,
+    },
+    {
+      icon,
+    },
     );
-})
+
+    adPin
+      .addTo(map)
+      .bindPopup(createPopupCard(item),
+        {
+          keepInView: true,
+        },
+      );
+  })
+}
+
+export {renderPins};
+
